requirements.txt
gradio
nbdev
Pillow
pdf2image